rename LOG_ERROR* macros to BREAK_ERROR*, remove unused code instrumentation
